Qt: Fix Linguist issues and format files
